+++ backend/src/main/java/site/kkokkio/domain/member/dto/TokenDto.java
@@ -1,0 +1,4 @@
+package site.kkokkio.domain.member.dto;
+
+public record TokenDto(String accessToken, String refreshToken) {
+}
