--- backend/src/main/java/site/kkokkio/domain/member/dto/TokenResponse.java
@@ -1,4 +1,0 @@
-package site.kkokkio.domain.member.dto;
-
-public record TokenResponse(String accessToken, String refreshToken) {
-}
