+++ backend/src/main/java/site/kkokkio/domain/member/service/AuthService.java
@@ -13,7 +13,7 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import site.kkokkio.domain.member.controller.dto.MemberLoginResponse;
-import site.kkokkio.domain.member.dto.TokenResponse;
+import site.kkokkio.domain.member.dto.TokenDto;
 import site.kkokkio.domain.member.entity.Member;
 import site.kkokkio.global.exception.ServiceException;
 import site.kkokkio.global.util.JwtUtils;
@@ -61,7 +61,7 @@ public class AuthService {
 
 	/** Refresh Token 검증 후 Access/Refresh 재발급 + Redis 갱신 + 쿠키 업데이트 */
 	@Transactional
-	public TokenResponse refreshToken(HttpServletRequest request, HttpServletResponse response) {
+	public TokenDto refreshToken(HttpServletRequest request, HttpServletResponse response) {
 		String rt = jwtUtils.getRefreshTokenFromCookies(request)
 			.orElseThrow(() -> new ServiceException("401-1", "리프레시 토큰이 없습니다."));
 
@@ -81,7 +81,7 @@ public class AuthService {
 		// 쿠키에도 업데이트
 		jwtUtils.setJwtInCookie(newAt, response);
 
-		return new TokenResponse(newAt, rt);
+		return new TokenDto(newAt, rt);
 	}
 
 	/** 로그아웃 -> Access Token 블랙리스트 등록 + Refresh Token 삭제 + 쿠키 삭제 */
