+++ backend/src/test/java/site/kkokkio/domain/member/service/MailServiceTest.java
@@ -23,7 +23,7 @@ import org.springframework.test.util.ReflectionTestUtils;
 
 import jakarta.mail.MessagingException;
 import jakarta.mail.internet.MimeMessage;
-import site.kkokkio.domain.member.dto.EmailVerificationRequest;
+import site.kkokkio.domain.member.controller.dto.EmailVerificationRequest;
 import site.kkokkio.domain.member.entity.Member;
 import site.kkokkio.domain.member.repository.MemberRepository;
